routes: migrate sessions-git + preview domains (13 routes)
Move to:
- src/server/routes/sessions-git.ts (10): /api/sessions/:id/file-content,
  git-status, git-diff, commits, pr-status, git-pull, git-push,
  git-squash-push, git-merge-primary, pr-merge
- src/server/routes/preview.ts (3): /api/preview/mount POST/GET,
  /api/sessions/:id/preview-events SSE

Drop now-unused imports (exec, shouldSkipRemotePush, previewMount,
broadcastPreviewChanged, getCachedPrStatus, batchGitStatus, execGit,
getGitDiff). Re-export of git-status test hooks reorganised into a
single `export … from` block.

Net: server.ts -703 lines (4136 → 3433). No behaviour change.
