+++ src/server/server.ts
@@ -1,4 +1,4 @@
-import { exec, execFile as execFileCb } from "node:child_process";
+import { execFile as execFileCb } from "node:child_process";
 import { promisify } from "node:util";
 import { randomUUID } from "node:crypto";
 import fs from "node:fs";
@@ -34,7 +34,7 @@ import { BgProcessManager } from "./agent/bg-process-manager.js";
 import { sessionFileRead, type SessionFsContext } from "./agent/session-fs.js";
 import { readTranscript, TranscriptReaderError } from "./agent/transcript-reader.js";
 
-import { isGitRepo, getRepoRoot, shouldSkipRemotePush, stripTokenFromGitUrl, detectPrimaryBranch } from "./skills/git.js";
+import { isGitRepo, getRepoRoot, stripTokenFromGitUrl, detectPrimaryBranch } from "./skills/git.js";
 import { VerificationHarness } from "./agent/verification-harness.js";
 import { validateAnswers, crossValidate, type UserQuestion } from "./agent/ask-user-choices-validation.js";
 import { buildAskResponseEnvelope, findAskResponseAnswers } from "../shared/ask-envelope.js";
@@ -65,8 +65,6 @@ import { CookieStore, issueIfMissing as issueCookieIfMissing, tryAuth as cookieT
 import { handlePreviewRequest } from "./preview/content-route.js";
 import { progressBus as searchProgressBus } from "./search/progress-bus.js";
 import { isSandboxAllowed } from "./auth/sandbox-guard.js";
-import * as previewMount from "./preview/mount.js";
-import { broadcastPreviewChanged, subscribePreviewChanged } from "./preview/events.js";
 import { startupAigwCheck, writeContextWindowOverrides } from "./agent/aigw-manager.js";
 import { writeOpenAIModelAdditions } from "./agent/openai-model-additions.js";
 import { ReviewAnnotationStore } from "./review-annotation-store.js";
@@ -97,28 +95,17 @@ import {
 	type ProposalType,
 } from "./proposals/proposal-files.js";
 
-const execAsync = promisify(exec);
 const execFileAsync = promisify(execFileCb);
 
 // ── Git helpers, status cache, PR cache, goal-branch deletion ──
 // All extracted to ./git/* — re-exported from this file so test hooks remain
 // importable from "./server.js" (see tests/e2e/git-status-caching.spec.ts).
-import { execGit, execGitSafe } from "./git/git-exec.js";
-import { getGitDiff } from "./git/git-diff.js";
-import { getCachedPrStatus } from "./git/pr-status.js";
+import { execGitSafe } from "./git/git-exec.js";
 import { deleteRemoteGoalBranches } from "./git/goal-branches.js";
 import { clearPrStatusCache } from "./git/pr-status.js";
-import {
-	batchGitStatus,
-	invalidateGitStatusCache,
-	__getGitStatusInvocationCount,
-	__resetGitStatusInvocationCount,
-	__setGitStatusFake,
-	__clearGitStatusFake,
-	__forceGitStatusCacheExpiry,
-	type GitStatusResult,
-} from "./git/git-status.js";
-export type { GitStatusResult };
+// Re-export git-status test hooks so existing imports from "./server.js" keep
+// working (see tests/e2e/git-status-caching.spec.ts).
+export type { GitStatusResult } from "./git/git-status.js";
 export {
 	invalidateGitStatusCache,
 	__getGitStatusInvocationCount,
@@ -126,7 +113,7 @@ export {
 	__setGitStatusFake,
 	__clearGitStatusFake,
 	__forceGitStatusCacheExpiry,
-};
+} from "./git/git-status.js";
 
 /** Cached Docker availability result to avoid running `docker info` per session creation */
 let _dockerAvailCache: { available: boolean; error?: string; ts: number } | null = null;
@@ -3123,108 +3110,6 @@ async function handleApiRoute(
 		return;
 	}
 
-	// GET /api/sessions/:id/file-content?path=<relative-or-absolute>&snapshotId=<id>
-	// Reads a text file for inline preview. When snapshotId is provided:
-	//   - If a snapshot exists on disk, returns the snapshot (historical state)
-	//   - Otherwise reads the live file and saves a snapshot for future refreshes
-	if (req.method === "GET" && url.pathname.startsWith("/api/sessions/") && url.pathname.endsWith("/file-content")) {
-		const id = url.pathname.split("/")[3];
-		const session = sessionManager.getSession(id);
-		if (!session) { json({ error: "Session not found" }, 404); return; }
-
-		const filePath = url.searchParams.get("path");
-		if (!filePath) { json({ error: "Missing path parameter" }, 400); return; }
-
-		const snapshotId = url.searchParams.get("snapshotId");
-		const snapshotDir = path.join(bobbitStateDir(), "html-snapshots");
-		const snapshotFile = snapshotId ? path.join(snapshotDir, `${snapshotId.replace(/[^a-zA-Z0-9_-]/g, "")}.html`) : null;
-
-		// Return existing snapshot if available
-		if (snapshotFile && fs.existsSync(snapshotFile)) {
-			try {
-				const content = fs.readFileSync(snapshotFile, "utf-8");
-				json({ content });
-			} catch {
-				json({ error: "Snapshot read failed" }, 500);
-			}
-			return;
-		}
-
-		// Read live file
-		const resolved = path.isAbsolute(filePath)
-			? path.resolve(filePath)
-			: path.resolve(session.cwd, filePath);
-
-		try {
-			const stat = fs.statSync(resolved);
-			if (stat.isDirectory() || stat.size > 512 * 1024) {
-				json({ error: "File too large or is a directory" }, 400);
-				return;
-			}
-			const content = fs.readFileSync(resolved, "utf-8");
-
-			// Save snapshot for future refreshes
-			if (snapshotFile) {
-				try {
-					fs.mkdirSync(snapshotDir, { recursive: true });
-					fs.writeFileSync(snapshotFile, content, "utf-8");
-				} catch { /* best-effort */ }
-			}
-
-			json({ content });
-		} catch {
-			json({ error: "File not found" }, 404);
-		}
-		return;
-	}
-
-	// GET /api/sessions/:id/git-status — get git status for session's working directory (async)
-	if (req.method === 'GET' && url.pathname.startsWith('/api/sessions/') && url.pathname.endsWith('/git-status')) {
-		const id = url.pathname.split('/')[3];
-		const session = sessionManager.getSession(id);
-		if (!session) {
-			json({ error: "Session not found" }, 404);
-			return;
-		}
-		const cwd = session.cwd;
-		const cid = session.sandboxed ? session.containerId : undefined;
-
-		if (!cid && !fs.existsSync(cwd)) { json({ error: "Working directory not found" }, 404); return; }
-
-		// Optional: run git fetch first when ?fetch=true is passed
-		const sessUntracked = url.searchParams.get('untracked') === '1';
-		if (url.searchParams.get('fetch') === 'true') {
-			try { await execGit('git fetch --quiet', cwd, 15000, cid); } catch { /* best-effort */ }
-			invalidateGitStatusCache(cwd, cid);
-		}
-
-		// Single attempt — native parallel execFile is fast (50–150 ms p50 on
-		// Windows) and errors are not cached, so the client retry loop in
-		// `git-status-refresh.ts` (4 attempts × 0/500/2000/5000 ms backoff) is
-		// the resilience layer for transient failures.
-		let result: Awaited<ReturnType<typeof batchGitStatus>> | undefined;
-		try {
-			result = await batchGitStatus(cwd, cid, { untracked: sessUntracked });
-		} catch (err: any) {
-			console.error("[git-status handler] error for session", id, "cwd=", cwd, "code=", err?.code, "signal=", err?.signal, "killed=", err?.killed, "stderr=", err?.stderr, "message=", err?.message);
-			jsonError(500, err, { error: err?.stderr?.trim() || err?.message || "git status failed" });
-			return;
-		}
-		if (!result) { json({ error: "Not a git repository" }, 400); return; }
-
-		json(result);
-
-		// Auto-push: for feature branches with unpushed commits, push in background
-		if (!shouldSkipRemotePush()) {
-			if (!result.isOnPrimary && result.ahead > 0 && result.hasUpstream) {
-				execAsync('git push', { cwd, encoding: "utf-8", timeout: 30000 }).catch(() => {});
-			} else if (!result.isOnPrimary && !result.hasUpstream && result.branch && /^session\//.test(result.branch)) {
-				// Session branches without upstream: set up tracking and push
-				execAsync(`git push -u origin ${result.branch}`, { cwd, encoding: "utf-8", timeout: 30000 }).catch(() => {});
-			}
-		}
-		return;
-	}
 	// GET /api/sessions/:id/tool-content/:messageIndex/:blockIndex — lazy-load full tool input content
 	const toolContentMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)\/tool-content\/(\d+)\/(\d+)$/);
 	if (toolContentMatch && req.method === "GET") {
@@ -3312,594 +3197,6 @@ async function handleApiRoute(
 				jsonError(500, err, { error: "internal_error", detail: String(err) });
 			}
 		}
-		return;
-	}
-
-	// GET /api/sessions/:id/git-diff — unified diff for session working directory
-	if (req.method === 'GET' && url.pathname.startsWith('/api/sessions/') && url.pathname.endsWith('/git-diff')) {
-		const id = url.pathname.split('/')[3];
-		const session = sessionManager.getSession(id);
-		if (!session) { json({ error: "Session not found" }, 404); return; }
-		const cwd = session.cwd;
-		const cid = session.sandboxed ? session.containerId : undefined;
-		if (!cid && !fs.existsSync(cwd)) { json({ error: "Working directory not found" }, 404); return; }
-		const file = url.searchParams.get("file") || undefined;
-		try {
-			const diff = await getGitDiff(cwd, file, cid);
-			json({ diff });
-		} catch (err: any) {
-			if (err.message === "INVALID_PATH") { json({ error: "Invalid file path" }, 400); return; }
-			if (err.message === "NO_DIFF") { json({ error: "No diff found" }, 404); return; }
-			jsonError(500, err);
-		}
-		return;
-	}
-	// GET /api/sessions/:id/commits — unpushed commits for session
-	if (req.method === 'GET' && url.pathname.startsWith('/api/sessions/') && url.pathname.endsWith('/commits')) {
-		const id = url.pathname.split('/')[3];
-		const session = sessionManager.getSession(id);
-		if (!session) { json({ error: 'Session not found' }, 404); return; }
-		const cwd = session.cwd;
-		const cid = session.sandboxed ? session.containerId : undefined;
-		if (!cid && !fs.existsSync(cwd)) { json({ commits: [] }); return; }
-		try {
-			let branch = '';
-			try { branch = await execGit('git rev-parse --abbrev-ref HEAD', cwd, 5000, cid); }
-			catch { json({ commits: [] }); return; }
-
-			let hasUpstream = false;
-			try { await execGit(`git rev-parse --abbrev-ref ${branch}@{u}`, cwd, 5000, cid); hasUpstream = true; } catch {}
-
-			const limit = 50;
-			const direction = url.searchParams.get('direction'); // 'behind' to show incoming commits
-			const vs = url.searchParams.get('vs'); // 'primary' to compare vs origin/master
-			let rangeSpec: string;
-			if (vs === 'primary') {
-				// Compare against origin/<primary>
-				let primaryBranch = 'master';
-				try {
-					const remoteHead = await execGit('git symbolic-ref refs/remotes/origin/HEAD', cwd, 5000, cid);
-					primaryBranch = remoteHead.replace('refs/remotes/origin/', '');
-				} catch {
-					try { await execGit('git rev-parse --verify refs/heads/master', cwd, 5000, cid); primaryBranch = 'master'; }
-					catch { try { await execGit('git rev-parse --verify refs/heads/main', cwd, 5000, cid); primaryBranch = 'main'; } catch {} }
-				}
-				let primaryRef = primaryBranch;
-				try { await execGit(`git rev-parse --verify origin/${primaryBranch}`, cwd, 5000, cid); primaryRef = `origin/${primaryBranch}`; } catch {}
-				rangeSpec = direction === 'behind' ? `HEAD..${primaryRef}` : `${primaryRef}..HEAD`;
-			} else {
-				rangeSpec = direction === 'behind' && hasUpstream
-					? 'HEAD..@{u}'
-					: hasUpstream ? '@{u}..HEAD' : `-${limit} HEAD`;
-			}
-
-			const out = await execGit(`git log --format="%H|%h|%s|%an|%aI" --shortstat ${rangeSpec}`, cwd, 10000, cid);
-			const lines = out.split('\n');
-			const commits: Array<{sha: string; shortSha: string; message: string; author: string; timestamp: string; filesChanged: number; insertions: number; deletions: number}> = [];
-
-			for (let i = 0; i < lines.length; i++) {
-				const line = lines[i];
-				if (!line.includes('|')) continue;
-				const parts = line.split('|');
-				if (parts.length < 5) continue;
-				const [sha, shortSha, message, author, timestamp] = parts;
-				// Next non-empty line should be the shortstat
-				let filesChanged = 0, insertions = 0, deletions = 0;
-				for (let j = i + 1; j < Math.min(i + 3, lines.length); j++) {
-					const statLine = lines[j].trim();
-					if (statLine.includes('file') && statLine.includes('changed')) {
-						const fm = statLine.match(/(\d+) file/);
-						const im = statLine.match(/(\d+) insertion/);
-						const dm = statLine.match(/(\d+) deletion/);
-						if (fm) filesChanged = parseInt(fm[1], 10);
-						if (im) insertions = parseInt(im[1], 10);
-						if (dm) deletions = parseInt(dm[1], 10);
-						break;
-					}
-				}
-				commits.push({ sha, shortSha, message, author, timestamp, filesChanged, insertions, deletions });
-			}
-
-			json({ commits });
-		} catch (e: any) {
-			json({ error: 'Failed to read git log', detail: e.message }, 500);
-		}
-		return;
-	}
-	// GET /api/sessions/:id/pr-status — PR status for session's branch
-	if (req.method === 'GET' && url.pathname.startsWith('/api/sessions/') && url.pathname.endsWith('/pr-status')) {
-		const id = url.pathname.split('/')[3];
-		const session = sessionManager.getSession(id);
-		if (!session) { json({ error: "Session not found" }, 404); return; }
-		const cwd = session.cwd;
-		const cid = session.sandboxed ? session.containerId : undefined;
-		if (!cid && !fs.existsSync(cwd)) { json({ error: "Working directory not found" }, 404); return; }
-		// Use goal branch if available so we find the right PR even if the worktree HEAD diverged.
-		// For non-goal sessions, fall back to the session's persisted branch — needed for sandbox
-		// sessions where the host worktree may not have the right branch checked out.
-		const goalBranch = session.goalId ? getGoalAcrossProjects(session.goalId)?.branch : undefined;
-		let sessionBranch = goalBranch || sessionManager.getPersistedSession(id)?.branch;
-		// For sandboxed sessions, the persisted branch may not match the actual container branch
-		// (e.g. gateway assigns a different worktree name). Detect the real branch from the container.
-		if (cid && cwd) {
-			try {
-				const actualBranch = await execGit("git rev-parse --abbrev-ref HEAD", cwd, 5000, cid);
-				if (actualBranch && actualBranch !== "HEAD") sessionBranch = actualBranch;
-			} catch { /* fall back to persisted branch */ }
-		}
-		// PR status uses `gh` CLI which needs host filesystem — use worktreePath for sandboxed sessions
-		const prCwd = cid ? (session.worktreePath || process.cwd()) : cwd;
-		const pr = await getCachedPrStatus(prCwd, sessionBranch, process.cwd());
-		if (pr) {
-			const goalId = session.goalId;
-			if (goalId) prStatusStore.set(goalId, pr);
-			json(pr);
-		} else { json({ error: "No PR found" }, 404); }
-		return;
-	}
-
-	// POST /api/sessions/:id/git-pull — pull latest from remote
-	if (req.method === 'POST' && url.pathname.startsWith('/api/sessions/') && url.pathname.endsWith('/git-pull')) {
-		const id = url.pathname.split('/')[3];
-		const session = sessionManager.getSession(id);
-		if (!session) { json({ error: "Session not found" }, 404); return; }
-		const cwd = session.cwd;
-		const cid = session.sandboxed ? session.containerId : undefined;
-		if (!cid && !fs.existsSync(cwd)) { json({ error: "Working directory not found" }, 404); return; }
-		try {
-			const output = await execGit('git pull', cwd, 30000, cid);
-			invalidateGitStatusCache(cwd, cid);
-			json({ ok: true, output });
-		} catch (err: unknown) {
-			const msg = err instanceof Error ? err.message : String(err);
-			json({ error: msg }, 500);
-		}
-		return;
-	}
-
-	// POST /api/sessions/:id/git-push — push local commits to remote
-	if (req.method === 'POST' && url.pathname.startsWith('/api/sessions/') && url.pathname.endsWith('/git-push')) {
-		if (shouldSkipRemotePush()) { json({ ok: true, output: "skipped (test mode)" }); return; }
-		const id = url.pathname.split('/')[3];
-		const session = sessionManager.getSession(id);
-		if (!session) { json({ error: "Session not found" }, 404); return; }
-		const cwd = session.cwd;
-		const cid = session.sandboxed ? session.containerId : undefined;
-		if (!cid && !fs.existsSync(cwd)) { json({ error: "Working directory not found" }, 404); return; }
-		try {
-			const output = await execGit('git push', cwd, 30000, cid);
-			invalidateGitStatusCache(cwd, cid);
-			json({ ok: true, output });
-		} catch (err: unknown) {
-			const msg = err instanceof Error ? err.message : String(err);
-			json({ error: msg }, 500);
-		}
-		return;
-	}
-
-	// POST /api/sessions/:id/git-squash-push — squash all branch commits and push directly to master
-	if (req.method === 'POST' && url.pathname.startsWith('/api/sessions/') && url.pathname.endsWith('/git-squash-push')) {
-		if (shouldSkipRemotePush()) { json({ ok: true, output: "skipped (test mode)" }); return; }
-		const id = url.pathname.split('/')[3];
-		const session = sessionManager.getSession(id);
-		if (!session) { json({ error: "Session not found" }, 404); return; }
-		const cwd = session.cwd;
-		const cid = session.sandboxed ? session.containerId : undefined;
-		if (!cid && !fs.existsSync(cwd)) { json({ error: "Working directory not found" }, 404); return; }
-		try {
-			// Detect primary branch
-			let primaryBranch = "master";
-			try {
-				const remoteHead = await execGit("git symbolic-ref refs/remotes/origin/HEAD", cwd, 5000, cid);
-				primaryBranch = remoteHead.replace("refs/remotes/origin/", "");
-			} catch {
-				try { await execGit("git rev-parse --verify refs/heads/master", cwd, 5000, cid); primaryBranch = "master"; }
-				catch { try { await execGit("git rev-parse --verify refs/heads/main", cwd, 5000, cid); primaryBranch = "main"; } catch { /* keep default */ } }
-			}
-
-			// Fetch latest master
-			await execGit(`git fetch origin ${primaryBranch}`, cwd, 30000, cid);
-			const primaryRef = `origin/${primaryBranch}`;
-
-			// Check we have commits ahead
-			const aheadCount = parseInt(await execGit(`git rev-list --count ${primaryRef}..HEAD`, cwd, 5000, cid), 10) || 0;
-			if (aheadCount === 0) { json({ error: "No commits ahead of master" }, 400); return; }
-
-			// Build commit message from branch commits
-			const logOutput = await execGit(`git log --format="%s" ${primaryRef}..HEAD`, cwd, 5000, cid);
-			const commitMessages = logOutput.trim().split("\n").filter(Boolean);
-			const branch = await execGit("git rev-parse --abbrev-ref HEAD", cwd, 5000, cid);
-			const summary = commitMessages.length === 1
-				? commitMessages[0]
-				: `Squash ${branch} (${commitMessages.length} commits)`;
-			const body = commitMessages.length > 1
-				? commitMessages.map(m => `- ${m}`).join("\n")
-				: "";
-			const fullMessage = body ? `${summary}\n\n${body}` : summary;
-
-			// Create squash commit on top of origin/master using plumbing (no checkout needed)
-			// 1. Create a tree that represents the merge result
-			const mergeTree = await execGit(`git merge-tree --write-tree ${primaryRef} HEAD`, cwd, 5000, cid);
-			// 2. Create a commit object with that tree, parented on origin/master
-			// For sandboxed sessions, write temp file inside container
-			const msgFile = cid ? `/tmp/SQUASH_MSG_${Date.now()}` : path.join(cwd, ".git", "SQUASH_MSG");
-			if (cid) {
-				await execFileAsync("docker", [
-					"exec", "-w", cwd, cid, "/bin/sh", "-c", `cat > ${msgFile} << 'BOBBIT_EOF'\n${fullMessage}\nBOBBIT_EOF`,
-				], { encoding: "utf-8", timeout: 5000, env: { ...process.env, MSYS_NO_PATHCONV: "1", MSYS2_ARG_CONV_EXCL: "*" } });
-			} else {
-				fs.writeFileSync(msgFile, fullMessage, "utf-8");
-			}
-			const squashCommit = await execGit(`git commit-tree ${mergeTree} -p ${primaryRef} -F "${msgFile}"`, cwd, 5000, cid);
-			if (cid) {
-				await execGit(`rm -f ${msgFile}`, cwd, 5000, cid).catch(() => {});
-			} else {
-				fs.unlinkSync(msgFile);
-			}
-			// 3. Push that commit to master
-			await execGit(`git push origin ${squashCommit}:refs/heads/${primaryBranch}`, cwd, 30000, cid);
-			invalidateGitStatusCache(cwd, cid);
-
-			json({ ok: true, output: `Squash pushed ${aheadCount} commit${aheadCount > 1 ? "s" : ""} to ${primaryBranch}` });
-		} catch (err: unknown) {
-			const msg = err instanceof Error ? err.message : String(err);
-			// Check for merge conflicts from merge-tree
-			if (msg.includes("CONFLICT") || msg.includes("merge-tree")) {
-				json({ error: "Merge conflicts with master. Use 'Merge master' first to resolve." }, 409);
-			} else {
-				json({ error: msg }, 500);
-			}
-		}
-		return;
-	}
-
-	// POST /api/sessions/:id/git-merge-primary — merge origin/master into current branch
-	if (req.method === 'POST' && url.pathname.startsWith('/api/sessions/') && url.pathname.endsWith('/git-merge-primary')) {
-		const id = url.pathname.split('/')[3];
-		const session = sessionManager.getSession(id);
-		if (!session) { json({ error: "Session not found" }, 404); return; }
-		const cwd = session.cwd;
-		const cid = session.sandboxed ? session.containerId : undefined;
-		if (!cid && !fs.existsSync(cwd)) { json({ error: "Working directory not found" }, 404); return; }
-		try {
-			// Detect primary branch
-			let primaryBranch = "master";
-			try {
-				const remoteHead = await execGit("git symbolic-ref refs/remotes/origin/HEAD", cwd, 5000, cid);
-				primaryBranch = remoteHead.replace("refs/remotes/origin/", "");
-			} catch {
-				try { await execGit("git rev-parse --verify refs/heads/master", cwd, 5000, cid); primaryBranch = "master"; }
-				catch { try { await execGit("git rev-parse --verify refs/heads/main", cwd, 5000, cid); primaryBranch = "main"; } catch { /* keep default */ } }
-			}
-			await execGit(`git fetch origin ${primaryBranch}`, cwd, 30000, cid);
-			const output = await execGit(`git rebase origin/${primaryBranch}`, cwd, 30000, cid);
-
-			// After rebase, check if orphaned commits remain (common after squash-merge PRs).
-			// If the tree is identical to origin/primary (no diff), the commits are redundant —
-			// reset to origin/primary to clean them up.
-			const aheadAfter = parseInt(await execGitSafe(`git rev-list --count origin/${primaryBranch}..HEAD`, cwd, "0", cid), 10) || 0;
-			if (aheadAfter > 0) {
-				const diff = await execGitSafe(`git diff origin/${primaryBranch}..HEAD`, cwd, "", cid);
-				if (diff.trim() === "") {
-					// Tree is identical — these are orphaned commits from a squash merge
-					await execGit(`git reset --hard origin/${primaryBranch}`, cwd, 10000, cid);
-					invalidateGitStatusCache(cwd, cid);
-					json({ ok: true, output: `Rebased and reset ${aheadAfter} orphaned commit(s) from squash merge` });
-					return;
-				}
-			}
-			invalidateGitStatusCache(cwd, cid);
-
-			json({ ok: true, output });
-		} catch (err: unknown) {
-			const msg = err instanceof Error ? err.message : String(err);
-			json({ error: msg }, 500);
-		}
-		return;
-	}
-
-	// POST /api/sessions/:id/pr-merge — merge PR for session's branch
-	if (req.method === 'POST' && url.pathname.startsWith('/api/sessions/') && url.pathname.endsWith('/pr-merge')) {
-		const id = url.pathname.split('/')[3];
-		const session = sessionManager.getSession(id);
-		if (!session) { json({ error: "Session not found" }, 404); return; }
-		const cwd = session.cwd;
-		const cid = session.sandboxed ? session.containerId : undefined;
-		if (!cid && !fs.existsSync(cwd)) { json({ error: "Working directory not found" }, 404); return; }
-		const body = await readBody(req);
-		const method = body?.method ?? "squash";
-		if (!["merge", "squash", "rebase"].includes(method)) {
-			json({ error: "Invalid merge method. Must be merge, squash, or rebase." }, 400);
-			return;
-		}
-		const sessAdminFlag = body?.admin ? " --admin" : "";
-		// Prefer the client-provided branch (headRefName from PR status) so the merge
-		// targets the exact PR the widget displayed — avoids mismatches when the session's
-		// persisted branch differs from the PR's head ref (e.g. staff/team agent worktrees).
-		const clientBranch = typeof body?.branch === "string" ? body.branch : undefined;
-		const goalBranch = session.goalId ? getGoalAcrossProjects(session.goalId)?.branch : undefined;
-		const sessMergeBranch = clientBranch || goalBranch || sessionManager.getPersistedSession(id)?.branch;
-		const sessMergeBranchArg = sessMergeBranch ? ` ${sessMergeBranch}` : "";
-		try {
-			// PR merge uses `gh` CLI — for sandboxed sessions, run on host worktree
-			const mergeCwd = cid ? (session.worktreePath || cwd) : cwd;
-			await execAsync(`gh pr merge${sessMergeBranchArg} --${method}${sessAdminFlag}`, { cwd: mergeCwd, encoding: "utf-8", timeout: 30000 });
-			clearPrStatusCache(cwd, sessMergeBranch);
-			json({ ok: true });
-		} catch (err: unknown) {
-			const msg = err instanceof Error ? err.message : String(err);
-			json({ error: msg }, 500);
-		}
-		return;
-	}
-
-	// ── Preview mount endpoints ──────────────────────────────────────
-	const VALID_SESSION_ID = /^[a-f0-9]{8}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{4}-[a-f0-9]{12}$/i;
-
-	// POST /api/preview/mount?sessionId=<sid> — v3 per-session preview mount.
-	// Accepts {html} (with optional {entry}) or {file: absolutePath}. Returns
-	// {url, path, entry, mtime}. See docs/design/embedded-html-preview-rewrite.md §6.
-	if (url.pathname === "/api/preview/mount" && req.method === "POST") {
-		const sessionId = url.searchParams.get("sessionId") || "";
-		if (!VALID_SESSION_ID.test(sessionId)) {
-			json({ error: "Invalid sessionId" }, 400);
-			return;
-		}
-		if (sandboxScope && !sandboxScope.sessionIds.has(sessionId)) {
-			json({ error: "Forbidden: session out of scope" }, 403);
-			return;
-		}
-		const body = await readBody(req).catch(() => ({}));
-		const hasHtml = typeof body?.html === "string";
-		const hasFile = typeof body?.file === "string" && body.file.length > 0;
-		const hasAssets = Array.isArray(body?.assets);
-		const hasManifest = typeof body?.manifest === "string" && body.manifest.length > 0;
-		if (!hasHtml && !hasFile) {
-			json({ error: "Body must contain one of 'html' or 'file'" }, 400);
-			return;
-		}
-		if (hasHtml && (hasAssets || hasManifest)) {
-			json({ error: "`assets`/`manifest` only valid with `file`" }, 400);
-			return;
-		}
-		try {
-			let result: previewMount.MountResult | previewMount.MountFileResult;
-			if (hasHtml) {
-				// `html` wins over `file` when both are provided.
-				let entry: string | undefined;
-				if (typeof body.entry === "string" && body.entry.length > 0) {
-					const e = body.entry;
-					if (e.includes("/") || e.includes("\\") || e.includes("..") || e.includes("\0")) {
-						json({ error: "Invalid entry name" }, 400);
-						return;
-					}
-					entry = e;
-				}
-				result = previewMount.writeInline(sessionId, body.html as string, entry);
-			} else {
-				const filePath = body.file as string;
-				if (!path.isAbsolute(filePath)) {
-					json({ error: "file path must be absolute" }, 400);
-					return;
-				}
-				if (!fs.existsSync(filePath)) {
-					json({ error: "file not found" }, 404);
-					return;
-				}
-				let stat: fs.Stats;
-				try { stat = fs.statSync(filePath); } catch {
-					json({ error: "file not found" }, 404);
-					return;
-				}
-				if (!stat.isFile()) {
-					json({ error: "path is not a regular file" }, 404);
-					return;
-				}
-				const base = path.basename(filePath).toLowerCase();
-				if (!base.endsWith(".html") && !base.endsWith(".htm")) {
-					json({ error: "file must end in .html or .htm" }, 400);
-					return;
-				}
-				// Collect assets from inline `assets[]` and optional `manifest` JSON.
-				const declared: string[] = [];
-				if (hasAssets) {
-					for (const a of body.assets as unknown[]) {
-						if (typeof a !== "string") {
-							json({ error: "`assets[]` entries must be strings" }, 400);
-							return;
-						}
-						declared.push(a);
-					}
-				}
-				if (hasManifest) {
-					const manifestRel = body.manifest as string;
-					if (path.isAbsolute(manifestRel) || manifestRel.includes("\0") ||
-						manifestRel.includes("\\") || manifestRel.split("/").some(s => s === "..")) {
-						json({ error: "Invalid manifest path" }, 400);
-						return;
-					}
-					const manifestAbs = path.resolve(path.dirname(filePath), manifestRel);
-					if (!fs.existsSync(manifestAbs)) {
-						json({ error: `Manifest '${manifestRel}' not found` }, 404);
-						return;
-					}
-					let manifestParsed: any;
-					try {
-						manifestParsed = JSON.parse(fs.readFileSync(manifestAbs, "utf-8"));
-					} catch (err: any) {
-						jsonError(400, err, { error: `Manifest JSON parse error: ${err?.message ?? err}` });
-						return;
-					}
-					if (!manifestParsed || !Array.isArray(manifestParsed.assets)) {
-						json({ error: "Manifest must be an object with an `assets[]` array" }, 400);
-						return;
-					}
-					for (const a of manifestParsed.assets) {
-						if (typeof a !== "string") {
-							json({ error: "Manifest `assets[]` entries must be strings" }, 400);
-							return;
-						}
-						declared.push(a);
-					}
-				}
-				// De-duplicate while preserving order.
-				const seen = new Set<string>();
-				const dedup: string[] = [];
-				for (const a of declared) {
-					const k = a.trim();
-					if (seen.has(k)) continue;
-					seen.add(k);
-					dedup.push(a);
-				}
-				result = previewMount.mountFile(sessionId, filePath, dedup);
-			}
-			broadcastPreviewChanged(sessionId, {
-				entry: result.entry,
-				mtime: result.mtime,
-				url: result.url,
-				path: result.path,
-			});
-			json(result);
-			return;
-		} catch (err: any) {
-			if (err && err instanceof previewMount.PreviewMountError) {
-				jsonError(err.statusCode, err);
-				return;
-			}
-			jsonError(500, err, { error: `preview mount failed: ${err?.message ?? String(err)}` });
-			return;
-		}
-	}
-
-	// GET /api/preview/mount?sessionId=<sid> — bootstrap the preview panel after
-	// session select. Returns the current entry/mtime/url/path for the mount,
-	// or 404 if the mount is empty / nonexistent. Same auth as the POST.
-	if (url.pathname === "/api/preview/mount" && req.method === "GET") {
-		const sessionId = url.searchParams.get("sessionId") || "";
-		if (!VALID_SESSION_ID.test(sessionId)) {
-			json({ error: "Invalid sessionId" }, 400);
-			return;
-		}
-		if (sandboxScope && !sandboxScope.sessionIds.has(sessionId)) {
-			json({ error: "Forbidden: session out of scope" }, 403);
-			return;
-		}
-		try {
-			const { pickEntry } = await import("./preview/content-route.js");
-			const dir = previewMount.mountDir(sessionId);
-			const entry = pickEntry(dir);
-			if (!entry) {
-				json({ error: "no preview mount" }, 404);
-				return;
-			}
-			const entryPath = path.join(dir, entry);
-			let stat: fs.Stats;
-			try { stat = fs.statSync(entryPath); } catch {
-				json({ error: "no preview mount" }, 404);
-				return;
-			}
-			json({
-				url: `/preview/${sessionId}/${entry}`,
-				path: entryPath,
-				entry,
-				mtime: Math.floor(stat.mtimeMs),
-			});
-			return;
-		} catch (err: any) {
-			if (err && err instanceof previewMount.PreviewMountError) {
-				jsonError(err.statusCode, err);
-				return;
-			}
-			jsonError(500, err, { error: `preview mount lookup failed: ${err?.message ?? String(err)}` });
-			return;
-		}
-	}
-
-	// GET /api/sessions/:sid/preview-events — SSE stream of preview-changed events
-	// for the per-session preview mount. Cookie auth (or admin bearer) only;
-	// sandbox tokens are not permitted (handled by the route-guard above).
-	const previewEventsMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)\/preview-events$/);
-	if (previewEventsMatch && req.method === "GET") {
-		const sid = previewEventsMatch[1];
-		if (!VALID_SESSION_ID.test(sid)) {
-			json({ error: "Invalid sessionId" }, 400);
-			return;
-		}
-		if (sandboxScope) {
-			json({ error: "Forbidden" }, 403);
-			return;
-		}
-		res.writeHead(200, {
-			"Content-Type": "text/event-stream",
-			"Cache-Control": "no-cache",
-			"Connection": "keep-alive",
-			"X-Accel-Buffering": "no",
-		});
-		try { (res as { flushHeaders?: () => void }).flushHeaders?.(); } catch { /* ok */ }
-		// Initial hello so the client knows the stream is live.
-		res.write(`event: hello\ndata: ${JSON.stringify({ ts: Date.now() })}\n\n`);
-
-		// Subscribe to the in-process preview-changed channel populated by the
-		// mount POST endpoint. Payload shape `{entry, mtime, url, path}` is
-		// forwarded verbatim — the client reads `entry` to seed the iframe.
-		const unsubscribe = subscribePreviewChanged(sid, payload => {
-			try {
-				res.write(`event: preview-changed\ndata: ${JSON.stringify(payload)}\n\n`);
-			} catch { /* socket closed */ }
-		});
-		// Bootstrap: if a mount already exists for this session, emit the
-		// current state synchronously so the just-connected client doesn't
-		// wait for the next agent write. Avoids a race where
-		// broadcastPreviewChanged fires between EventSource open and the
-		// subscription being registered. Payload shape `{entry, mtime, url,
-		// path}` matches broadcastPreviewChanged so the client doesn't need
-		// to distinguish bootstrap from live events.
-		try {
-			const { pickEntry } = await import("./preview/content-route.js");
-			const dir = previewMount.mountDir(sid);
-			if (fs.existsSync(dir)) {
-				const entry = pickEntry(dir);
-				if (entry) {
-					const entryPath = path.join(dir, entry);
-					const stat = fs.statSync(entryPath);
-					res.write(`event: preview-changed\ndata: ${JSON.stringify({
-						entry,
-						mtime: Math.floor(stat.mtimeMs),
-						url: `/preview/${sid}/${entry}`,
-						path: entryPath,
-					})}\n\n`);
-				}
-			}
-		} catch { /* ok — bootstrap is best-effort */ }
-		const keepalive = setInterval(() => {
-			try { res.write(":keepalive\n\n"); } catch { /* ok */ }
-		}, 25_000);
-		if (typeof keepalive.unref === "function") keepalive.unref();
-		const cleanup = () => {
-			clearInterval(keepalive);
-			try { unsubscribe(); } catch { /* ok */ }
-		};
-		req.on("close", cleanup);
-		req.on("error", cleanup);
-		return;
-	}
-
-	// ── Draft endpoints ─────────────────────────────────────────────
-
-	// PUT|POST /api/sessions/:id/draft — upsert a draft
-	// POST is accepted alongside PUT because navigator.sendBeacon (used for
-	// beforeunload draft flush) always sends POST requests.
-	const draftPutMatch = url.pathname.match(/^\/api\/sessions\/([^/]+)\/draft$/);
-	if (draftPutMatch && (req.method === "PUT" || req.method === "POST")) {
-		const id = draftPutMatch[1];
-		const body = await readBody(req);
-		if (!body || typeof body.type !== "string") {
-			json({ error: "Missing type" }, 400);
-			return;
-		}
-		const ok = sessionManager.setDraft(id, body.type, body.data);
-		if (!ok) { json({ error: "Session not found" }, 404); return; }
-		json({ ok: true });
 		return;
 	}
